IS-1932: Move api-DTOs to api-folder
